hello_world/requirements.txt
requests
PyYAML==6.0
boto3